Add Video Mgr

+++ build.py
@@ -1,3 +1,4 @@
+from genericpath import isdir
 import os
 import shutil, errno
 
@@ -6,7 +7,8 @@ from flask_script import Manager
 from model.model2python import GenCodeCommand
 
 if __name__ == "__main__":
-    shutil.rmtree( "app")
+    if ( os.path.isdir("app")) : 
+        shutil.rmtree( "app")
     shutil.copytree( "skeleton", "app")
     print (  os.path.dirname(os.path.realpath(__file__)))
     r = GenCodeCommand()
